feature: modifying result type to long for grather solution values

+++ src/Fibonacci.java
@@ -1,19 +1,19 @@
 public class Fibonacci {
 
-    public static int recursiveSolution(int n){
+    public static long recursiveSolution(int n){
         if(n < 0) return -1;
         if(n <= 1) return n;
 
         return recursiveSolution(n -1) + recursiveSolution(n - 2);
     }
 
-    public static int linearSolution(int n){
+    public static long linearSolution(int n){
         if(n < 0) return -1;
         if(n <= 1) return n;
 
-        int last = 0;
-        int current = 1;
-        int solution = 1;
+        long last = 0;
+        long current = 1;
+        long solution = 1;
 
         for(int i = 2; i <= n; i++){
             solution = current + last;
@@ -24,7 +24,7 @@ public class Fibonacci {
         return solution;
     }
 
-    public static void printResult(int n, int result){
+    public static void printResult(int n, long result){
         if(result < 0) System.out.println("Valor inválido");
         else System.out.println("O "+n+"º elemento da sequencia de fibonacci é "+ result);
     }
